% Auto-generated: do not edit by hand
\name{sylvereyeRoadNetwork}

\alias{sylvereyeRoadNetwork}

\title{SylvereyeRoadNetwork component}

\description{

}

\usage{
sylvereyeRoadNetwork(children=NULL, id=NULL, clicked_edge=NULL,
clicked_marker=NULL, clicked_node=NULL, current_state=NULL,
debug_options=NULL, edge_options=NULL, edges_data=NULL,
map_center=NULL, map_max_zoom=NULL, map_min_zoom=NULL,
map_style=NULL, map_zoom=NULL, marker_options=NULL,
markers_data=NULL, node_options=NULL, nodes_data=NULL,
pane=NULL, show_arrows=NULL, show_edges=NULL,
show_markers=NULL, show_nodes=NULL,
tile_layer_attribution=NULL, tile_layer_opacity=NULL,
tile_layer_subdomains=NULL, tile_layer_url=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{clicked_edge}{Named list. Clicked edge}

\item{clicked_marker}{Named list. Clicked marker}

\item{clicked_node}{Named list. Clicked node}

\item{current_state}{Named list. Current state}

\item{debug_options}{Named list. Debug options}

\item{edge_options}{Named list. Edge options}

\item{edges_data}{Unnamed list. Edges data}

\item{map_center}{Unnamed list. Map center}

\item{map_max_zoom}{Numeric. Map max zoom}

\item{map_min_zoom}{Numeric. Map min zoom}

\item{map_style}{Named list. Map style}

\item{map_zoom}{Numeric. Map zoom}

\item{marker_options}{Named list. Marker options}

\item{markers_data}{Unnamed list. Markers data}

\item{node_options}{Named list. Node options}

\item{nodes_data}{Unnamed list. Nodes data}

\item{pane}{Character. The leaflet pane of the component}

\item{show_arrows}{Logical. Show/hide arrows}

\item{show_edges}{Logical. Show/hide edges}

\item{show_markers}{Logical. Show/hide markers}

\item{show_nodes}{Logical. Show/hide nodes}

\item{tile_layer_attribution}{Character. Tile layer attribution}

\item{tile_layer_opacity}{Numeric. Tile layer opacity}

\item{tile_layer_subdomains}{Character. Tile layer subdomains}

\item{tile_layer_url}{Character. Tile layer URL}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

